style: move static variable out of function
fix: optimize pubkey generator

fix: change omnilock hash type

style: remove unused

fix: simplify h256 convert function

fix: remove redundant import
