+++ tx-builder/src/ckb/utils/mod.rs
@@ -1,3 +1,3 @@
 pub mod calc_amount;
 pub mod cell_data;
-pub mod script_utils;
+pub mod script;
